Cast total to double before dividing

+++ src/main/java/teammates/common/datatransfer/FeedbackConstantSumQuestionDetails.java
@@ -523,7 +523,7 @@ public class FeedbackConstantSumQuestionDetails extends FeedbackQuestionDetails 
         DecimalFormat df = new DecimalFormat("#.##");
         
         for(Entry<String, Integer[]> entry : optionTotalCount.entrySet() ){
-            double average = entry.getValue()[0]/entry.getValue()[1];
+            double average = (double)entry.getValue()[0]/entry.getValue()[1];
             fragments += entry.getKey() + ","
                       + df.format(average) + Const.EOL;
         }
